冗長なintキャストを除去する
\(int\)(?=int\(std::size)
7か所一致

s/\(int\)(?=int\(std::size)//g

+++ sakura_core/dlg/CDlgOpenFile_CommonFileDialog.cpp
@@ -227,7 +227,7 @@ UINT_PTR CALLBACK OFNHookProc(
 		L"LF (UNIX)",
 		L"CR (Mac)",
 	};
-	int nEolNameArrNum = (int)int(std::size(pEolNameArr));
+	int nEolNameArrNum = int(std::size(pEolNameArr));
 
 //	To Here	Feb. 9, 2001 genta
 	int	nRightMargin = 24;
