+++ sakura_core/env/CShareData.cpp
@@ -1463,7 +1463,7 @@ std::vector<STypeConfig*>& CShareData::GetTypeSettings()
 void CShareData::InitFileTree( SFileTree* setting )
 {
 	setting->m_bProject = true;
-	for(int i = 0; i < (int)int(std::size(setting->m_aItems)); i++){
+	for(int i = 0; i < int(std::size(setting->m_aItems)); i++){
 		SFileTreeItem& item = setting->m_aItems[i];
 		item.m_eFileTreeItemType = EFileTreeItemType_Grep;
 		item.m_szTargetPath = L"";
